requirements.txt
allennlp==0.9.0
boto3==1.17.101
cachetools==4.2.2
japanize_matplotlib==1.1.3
jsonlines==2.0.0
matplotlib==3.4.2
nltk==3.6.5
numpy==1.21.0
pandas==1.3.0
pcfg==0.1.5
plotly==5.3.1
python_dateutil==2.8.2
pytorch_pretrained_bert==0.6.2
scikit_learn==1.2.0
scipy==1.7.0
seaborn==0.11.1
sympy==1.10.1
torch==1.2.0
torchinfo==1.5.3
tqdm==4.61.1
transformers==4.25.1
transformers.egg==info
ujson==4.0.2
umap==0.1.1
umap_learn==0.5.1
